rag/basic/requirements.txt
langchain-community
langchain-text-splitters
langchain-ollama
langchain-chroma
pypdf
chromadb
pyyaml
